Dependencies: Update requirement `pymatgen>=2022.1.20`
The CI workflows required a separate installation step of `pymatgen`
before installing everything else. This is because older version of
`pymatgen` did not specify an explicit compatible version of `numpy` in
their build requirements. This made it possible that `pymatgen` got
built with an older version of `numpy` that later would get update when
the requirements of `aiida-core` got processed.

As of `pymatge==2022.1.20`, the build process properly specifies the
latest compatible numpy version, so this problem no longer occurs and
the custom installation step of `numpy` plus `pymatgen` can be removed.
See: https://github.com/materialsproject/pymatgen/commit/
ca664f142160a3e7a9f05c4ab4899f841e5bda3f  for the commit that fixed the
problem.

By removing this step, we can now also add support for Python 3.11 which
requires a more recent version than `1.21.4` which was being pinned for
the custom install step.