ORM: Remove `pymatgen` version check in `StructureData.set_pymatgen_structure` (#5777)
The `StructureData.set_pymatgen_structure` method performed an explicit
check on the version of `pymatgen` that is installed to determine whether
the `pymatgen.Site.specie` property could be used. However, this was
introduced in `pymatgen==2019.3.13` but the minimum required version
specified in the `atomic_tools` dependencies is `2019.7.2` so the check
is superfluous.