`PsqlDosMigrator`: refactor the connection handling
The `PsqlDosMigrator` requires a connection to the database for various
operations. To this end, it created an `Engine` from the connection
parameters provided by the `Profile`, through which a connection was
opened in the `_connection_context`. However, this connection was never
explicitly closed. Moreover, although the method allowed passing an
optional connection that was already open, this was not consistently
used in the entire class. This could lead to too many connections being
opened to the client resulting in an exception when the limit was
reached.

Here the class is refactored as follows:

 * The engine is created in the constructor and assigned to the private
   attribute `_engine`.
 * A `connection` property is added that will return an open connection.
   If the engine hasn't already been constructed or no connection is
   already open, the engine is constructed and a connection is opened
   which is then assigned to the `_connection` attribute.
 * The `close` method is added which when called will close the
   connection if it was opened and dispose of the engine. This should be
   called if the caller is done using the migrator and its connections
   to the database should be cleaned up.
 * The `_connection_context` context manager is removed and all internal
   methods now simply call the `connection` property to get an open
   connection.

The `PsqlDosBackend` uses the new connection handling interface of the
`PsqlDosMigrator` through the `migrator_context` context manager. Each
method that requires a migrator should request it through this method.
At the end of the context the `close` method is automatically called on
the migrator ensuring its opened connections are properly relinquished.