Fixtures: Add `aiida_computer_local` and `aiida_computer_ssh`
These fixtures return an instance of `Computer` that represents the
localhost computer with `core.local` and `core.ssh` as the transport
type, respectively. By default, the computers are also configured. The
`aiida_computer_ssh` will even create an SSH key pair automatically for
the session such that the SSH transport can actually be opened.

The fixtures both call the more generic `aiida_computer` which returns a
factory that gives greater control on how the computer is created and
configured. Usuallly the two wrappers are all that is needed for the
majority of tests.

The fixtures are placed in `pytest_fixtures` instead of the `conftest.py`
of `aiida-core`, because `aiida_computer` and `aiida_computer_local`
should be available to plugins. The `aiida_computer_ssh` is probably
mostly useful for tests in `aiida-core`, but for clarity it is kept
together with `aiida_computer_local`.

The existing fixture `aiida_localhost` now simply uses the
`aiida_computer_local` fixture as the implementation.