`InstalledCode`: Fix bug in `validate_filepath_executable` for SSH (#5787)
If the code was attached to a `Computer` with `core.ssh` transport, the
`validate_filepath_executable` would always return that the computer
could not be connected to. This is because the actual exception that was
caught does not come from `Computer.get_transport` but is a `TypeError`
raised by `transport.isfile` since it receives a `pathlib.Path` were it
expects a `str`.

The problem went undetected since it only tested against a computer with
`core.local` transport. A new test is added to also test against a
computer with `core.ssh` transport`.