`Process`: Store JSON-serializable metadata inputs on the node
The promise of AiiDA's provenance is that all inputs to a `Process` are
stored as nodes in the provenance graph linked to a node representing
the process. From the very early beginning, however, there needed to be
exceptions of inputs to processes that were not nodes. Notable examples
were the various "options" set for calculation jobs. Even in AiiDA v0.x
these "settings" as they were called back then, were stored in the
attributes of the node.

In the AiiDA v1.0 redesign, where all inputs of a process are defined
through the process spec, these non-node inputs were implemented by
allowing input ports to be made non-database storable, indicated by the
`non-db` argument in the port declaration. These inputs would be passed
to the `Process` instance and would be available during its lifetime,
but would not be stored in the database. Once again there are exceptions
as certain inputs defined by `aiida-core` are stored on the node, but in
various places. Notable examples are the `label` and `description` of
the process, and the `metadata.options` of the `CalcJob` class.

This historical decision has as a direct result in that it is difficult
if not impossible in certain cases to reconstruct the exact input
dictionary that was used to run a `Process` from the data stored on the
`ProcessNode`. From a provenance point of view, this is a huge weakpoint
and is what is being corrected here.

The input ports marked `non_db=True` on the base process classes
provided by `aiida-core`, `Process` and `CalcJob` were changed to use the
new `is_metadata` keyword instead in the previous commit, to remove the
inconsistency between the naming and behavior. In this commit, all inputs
that correspond to `is_metadata` ports and are JSON-serializable are stored
in the attributes of the process node under the key `metadata_inputs`.
All `is_metadata` input ports that are defined on process base classes by
`aiida-core`, such as `Process` and `CalcJob` *are* JSON serializable.
However, plugin packages can implement process classes with ports that
accept inputs that are not JSON serializable, which is why this
additional condition has to be added. But all inputs defined by
`aiida-core` should be covered.