`ProcessBuilder`: Fix bug in `_recursive_merge`
The `_recursive_merge` method could raise a `KeyError` for a process
builder that contains a dynamic port namespace. In this case it would be
possible to merge in a value that contains a nested dictionay which
would try to access directly `dictionary[key]`, where `dictionary` is
the namespace of the builder, and so hit a `KeyError` since this nested
namespace didn't exist yet in the dynamical namespace.

The solution is to explicitly check if the `key` already exists in the
builder's namespace, and if not, we assig the entire `value` to it, as
there is no nothing to recurse into.