CLI: Add short option names for `verdi code create` (#5799)
The `verdi code create` command replaced `verdi code setup` but unlike
its predecessor it only has long option names. To provide short options
as well, the `DynamicEntryPointCommandGroup` is extended to support the
`short_name` key in an option specification. If specified, the option
will provide this short option as well as the automatically determined
long option.

The `InstalledCode`, `PortableCode` and `ContainerizedCode` are updated
to provide short options for those that used to have one in the replaced
`verdi code setup` command.