`verdi run`: Do not add `pathlib.Path` instance to `sys.path`
The `verdi run` command updates `sys.path` to include the current
working directory such that modules in it can be imported by the script
being run. However, it was using `pathlib` to get the CWD and was adding
a `Path` instance to the path, which should only contain `str`
instances. This was uncovered by trying to import `tensorflow` which
also manipulates the `sys.path` and was raising an error since it was
trying to iterate over all elements and `pathlib.Path` is not iterable,
unlike `str`.