Docs: Fix the build on ReadTheDocs
There was a backtick missing in the `help` attribute of an input port
declaration of the `CalcJob` class. This class is rendered automatically
through our sphinx plugin in the `topics/calculations/usage.rst` file in
the "Options" section. No idea why this didn't fail before this because
this mistake was not introduced in this PR.

Also add config option for `myst-nb` to `conf.py` so that it shows the
stacktrace if an exception is raised during notebook execution.