`Process`: Update outputs before updating node process state
In `a7be795a26aea9fdff1d6f762582f1b77801ff3d` the order of the two
method calls in `Process.update_node_state` was reversed. The idea was
that the `update_outputs` call can except, for example if a `Process`
implementation adds an invalid output by calling `Process.out`. In this
case the process will go into the excepted state and the process state
would sometimes not be updated on the node. By reversing the two actions
the process state was hoped to be more accurate in these cases.

However, since this change, the `nightly` workflow started failing where
nested workchains would except because a subprocess didn't have an
output that they expected. It is not clear at all what the relation is
but it seems that reverting the change described above resolves the
problem. Unfortunately, the bug does not seem to be deterministic and
doesn't always manifest.

Here, we reverse the order to the original state, updating the outputs
first, but in order to guarantee that the process state is always
updated, even if the output update excepts, it is called in the finally
clause of the try-except in which the `update_outputs` is wrapped.