`Repository`: Cache the calculation of the hash
The `Repository.hash` method computes the hash of the repository by
iterating over all the objects it contains and creating the cumulative
hash of all their hashes put together. For large repositories this can
be a costly operation and, unless the content is changed, the value
should not change between invocations.

Therefore, the hash is stored as an instance attribute such that it can
be returned the next time it is requested instead of recomputing. The
hash is invalidated, by resetting it to `None` as soon as the state of
the repository is mutated.

To measure the impact on the performance, the benchmark script of the
docs `docs/source/howto/include/scripts/performance_benchmark_base.py`
was run against the `main` branch and this `fix` branch. The script was
launched with the `--without-daemon` option and was requested to run 50
`ArithmeticAddCalculation` in sequence. At the end, the script reports
the number of seconds taken per process. The benchmark was run on a
machine with a AMD Ryzen 5 3600 6-Core processor and an SSD for storage
using two different storage backends: the `PsqlDosBackend` and the
`PsqlAzureBlobStorage` of the `aiida-s3` package.

The latter swaps the disk object store on the local filesystem for a
container on the Azure Blob Storage service. This is done to amplify the
effect of the caching. For the disk object store on the local file
system, file access is very fast and so the caching of the repository
hash is negligible. However, if the repository access becomes slower,
such as with the object store, the effect becomes more pronounced.

The results of the benchmark are in the table below, where the number
reported is the time in seconds per process:

Branch | `core.psql_dos` | `s3.psql_azure_blob`
-------|-----------------|---------------------
`main` |   0.28          |  2.50
`fix`  |   0.28          |  1.95

As explained before, the effect of the change for the disk object store
on the local filesystem is negligible. However, for the blob storage the
run time is reduced by 20%.

As a final note, the absolute time of the `s3.psql_azure_blob` depends a
lot on the configuration of the blob container. For example, containers
in servers that are far away will naturally have a large overhead due to
the latency for each request. Same goes for containers that are
optimized for long term storage which will have larger request time for
retrievals. The absolute value is therefore not interesting as it can
vary a lot, but what is interesting is the difference in time between
the code that includes caching and that which doesn't.