`SqliteTempBackend`: Fix the `transaction` method
When trying to store nodes, for example by running a calcfunction,
SqlAlchemy would throw an exception complaining that the transaction had
already been closed and the context manager had to be closed first. The
implementation is fixed by matching it with that of the `PsqlDosBackend`
storage backend.

The error was only discovered through the execution of the `tutorial.md`
notebook in the documentation. An explicit unit test is added to make it
more obvious and easier to debug in case a regression is introduced.