`PsqlDosBackend`: Remove `disable_expire_on_commit` utility
The implementation of the backend `Group` ORM class for the
`PsqlDosBackend` used the `disable_expire_on_commit` context manager in
the `add_nodes` and `remove_nodes` methods. By default, the engine used
sets `expire_on_commit=True`, which means that all database model
instances in the session are expired once the session is committed,
which forces the instances to be refreshed from the database the next
time they are accessed.

This is a sensible and important setting, as it ensures that we are in
sync with the database when multiple processes are committing to the
database. However, expiring instances comes at a cost because they will
have to be refreshed at some point. It is unknown exactly why his change
was originally applied, but it is probably because expiring all nodes in
the session upon addition/removal of nodes from a group affected
performance badly and the probability of inconsistencies by omitting it
are less problematic.

The behavior was never tested though and with the recent refactor of the
transaction control in the `PsqlDosBackend` it seems the utility to
explicitly disable the expire on commit no longer seems necessary. In
the previous commit a test was added to test the utility correctly being
applied, but it works whether the utility is used or not at all. This
suggests that with the current proper use of a nested transaction in
`add_nodes` and `remove_nodes`, the nodes in the outer session are no
longer expired, making the utility obsolete and therefore it is removed.