Add test for `SqlaGroup.add_nodes` to test `disable_expire_on_commit`
The test passes, but it seems independent of the use of the utility
`disable_expire_on_commit`. It looks like this is because now a
transaction is properly used to add or remove nodes, and so the
instances in the parent session are not expired. It seems that the
proper use of transactions essentially makes `disable_expire_on_commit`
obsolete. When the same test is run on `v2.1`, where transactions were
not properly used, removing the context manager fails the test.