Process functions: Restore support for dynamic nested input namespaces (#5808)
This behavior was supported before v2.0.4 but was accidentally removed by
putting more strict type validation on inputs passed to process
functions. In d2323d462a9c84ba83f24117fb78e5b102a53e85, a bug was fixed
where it was possible to pass non-storable inputs to dynamic port
namespaces. For example, the following was allowed:

    @calcfunction
    def sum(**kwargs):
        return sum(kwargs.values())

    sum(**{'a': 1, 'b': 2})

However, since the values in the `kwargs` are not `Data` nodes (and this
was before automatic input serialization was added for process functions)
and so they could not be stored in the provenance graph. This would
therefore cause accidental provenance loss.

This bug was fixed in the aforementioned commit by explicitly setting
`valid_type=Data` on the inputs namespace of the `Process` base class.

The fix, however, inadvertently also removed support for arbitrarily
nested keyword arguments in process functions. While this would work
before, it would now raise a `ValidationError` saying that the dynamic
namespace received a `dict` instead of a `Data` node. The cause is that
the `PortNamespace.validate_dynamic_ports` did not recurse down nested
namespaces, but only validated top-level values.

This behavior is fixed in `plumpy==0.21.3` where now the validation
method will recursively go down any nested namespaces for a dynamic port
namespace and validate as long as all leaf-values match the `valid_type`
of the `PortNamespace`.