`Port`: Add the `is_metadata` keyword
In the engine redesign from `v0.x` to `v1.0` the process interface
needed a way to distinguish inputs that are lined up to the process node
as nodes themselves and those that are stored directly on the process
node, for example as an attribute. To this end the `non_db` keyword was
added to the `InputPort`. This keyword was used for the `metadata` input
namespace, to designate that these inputs would not be `Data` instances.

The name is quite unfortunate, however, since the inputs of these ports
actually do get stored in the database, contrary to what the keyword
suggests. The most straightforward solution would be to rename the
keyword, however, for better or worse, the keyword has been adopted by
(a limited amount of) plugin packages. A known application is to pass in
`Group` instances which are not storable as nodes, but can be passed
through a `non_db` port. Hypothetically, users may have used the port as
well to pass in sensitive data that should never be stored. Renaming the
port or changing its behavior is therefore likely to break existing
code.

Instead, the `is_metadata` keyword is added to the `InputPort` and
`PortNamespace` through the `WithMetadata` mixin. When set to `True`
this keyword functions as the original intention of the `non_db` flag
and an `is_metadata` port signals that its value will be stored in the
database but directly on the `ProcessNode` instead of being linked as a
`Data` node. The naming makes sense, because the only use of this
keyword is by the `metadata` input namespace that all `Process` classes
have. The inputs in this namespace are stored, through custom logic in
the `Process` and `CalcJob` class, in the attributes of the node or
dedicated columns of the node database model, such as the label and the
description.

This addition leaves the behavior of `non_db` inputs unchanged and so
plugin packages that use this keyword should continue to function as
before.