CLI: Add the `verdi calcjob remotecat` command (#4861)
The command can be used to inspect the files in the working directory
of a `CalcJob` while it is still running. The `outputcat` only works on
the files in the `retrieved` folder so only works for retrieved jobs.
There is the `verdi data core.remote cat` command, but there the user
needs to specify the identifier of the `remote_folder` of the calcjob
and the output file. The `verdi calcjob remotecat` command simplifies
these steps.

It was discussed to add the `--monitor` option to simulate the
functionality of a `tail -f` call on an output file, but this required
the adaptation of the `Transport` interface which may get refactored
soon, so it was decided to omit that feature for the time being.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>