Docs: Add `--without-daemon` flag to benchmark script
The benchmark script sends a number of `ArithmeticAddCalculation`s to
the daemon and reports the time taken per process at the end. This is
intended for users to run to get an idea of the performance of their
setup. However, running over the daemon is not necessary the most
straightforward when it comes to analysing the results and the timings
can differ quite a bit. It makes more sense to simply run the N jobs
serially, which should give a better more consistent picture of the
average overhead per process.