Tests: Use `use_subprocess=False` and `suppress_warnings=True`
Certain command line tests cannot be run with `use_subprocess=True`.
Examples are tests that use the `monkeypatch` fixture to patch certain
resources. If the command were to be executed in a subprocess, the patch
would no longer apply since it is only active in the current interpreter.
For those tests, the `use_subprocess` keyword of `run_cli_command` needs
to be set to `False`.

Also test that use the `empty_config` or `config_with_profile_factory`
fixtures need to run with `use_subprocess=False` since these fixtures
generate a temporary config that would not be used by the subprocess
since that uses the config at the `AIIDA_PATH` environment variable.

Then there are tests whose assertions very strictly check the output. If
the command is executed through a subprocess, any deprecation warnings
that are hit will be printed and the test might fail. This is not a
problem when running through the test runner because then the
deprecation warning will already have been emitted when the test suite
loaded and won't be fired another time when the test code is executed.

For this purpose the `suppress_warnings` argument is added to the
`run_cli_command` fixture. False by default, but when set to `True` the
`PYTHONWARNINGS` environment variable is set to `ignore::Warning` which
should cause all warnings to be suppressed.