Tests: Fix bugs revealed by running with `use_subprocess=True`
The addition of `use_subprocess=True` to the `run_cli_command` fixture
to actually run the command as a subprocess instead of through the test
runner in the same interpreter, revealed a number of small bugs in the
tests themselves that would cause them to fail. They are fixed such that
the tests pass regardless of the value of `use_subprocess`.