Fixtures: Add argument `use_subprocess` to `run_cli_command`
By default the `run_cli_command` fixture will run the provided command
through the `CliRunner` utility provided by `click`. However, this
suffers from a problem as the code pathway followed through the runner
is not identical as when the command is actually called through the CLI.

The reason is that in `click` commands do not have a reference to their
parents, and so when a subcommand is invoked through the runner, only
the code of the subcommand is executed. Any code that is part of its
parents, such as parameter validation and callbacks are skipped. In
addition, the runner is executed in the same interpreter as where the
test suite is being run, whereas a command invoked through the actual
CLI runs in a standalone interpreter. This may hide subtle bugs.

Therefore, the `use_subprocess` argument is added, which is `False` by
default, which when set to `True`, the command will not be called through
the runner, but as an actual subprocess using `subprocess.run`. This
guarantees that the tested pathway is identical to what an actual
invocation through the command line would be. It is not enabled by
default since running through a subprocess is slower and not always
necessary.