`verdi computer test`: Add test for login shell being slow
By default computers are configured to use a login shell, setting the
`use_login_shell` configuration option to `True`, which ensures that the
`Transport` always uses the `-l` option for bash when executing a
command. This can be important for certain compute environments where
the login script contains commands that are necessary for calculation
jobs to execute successfully.

However, this can incur significant slowdowns if the login script loading
is slow. This will affect the throughput of calculation jobs
significantly the source of which can be difficult to track down. That
is why a test is added to `verdi computer test`. It will execute
`whoami` over the transport with and without a login shell and if the
ratio exceeds 2, a warning is emitted, suggesting to turn off the option
unless strictly necessary.