CLI: Fix results not being committed in `verdi code delete`
The `verdi code delete` command would report that the `Code` was deleted
but the changes were actually not committed. This went unnoticed since
the test for it used the `click.CliRunner` which doesn't follow the
exact same path as a real invocation on the command line. In particular,
in this example the treatment of the session is different with it not
being committed properly in the command line invocation pathway. The
problem is shown by using `use_subprocess=True` when calling the command
using the `run_cli_command` fixture, after which the test fails.