`Scheduler`: Allow terminating job if submission script is invalid
Currently, when a `CalcJob` is launched, the job will be submitted to
the scheduler by calling `Scheduler.submit_from_script`. This is either
expected to succeed, and return the job id in that case, or to raise an
exception if the submission failed. If an exception is raised, the
exponential backoff retry mechanism kicks in, which will try again,
until the maximum number of retries is reached, at which point the job
will be paused.

This behavior is useful if the cause of the exception being raised is
transient, for example the scheduler being overloaded or the connection
to the remote computer failing. But for problems that will always fail,
this will just cause the job to be stuck as soon as it gets paused.
There should be a way for scheduler plugins to detect these terminal
problems, such as an invalid submission script, and communicate that the
calculation job should simply be terminated.

The signature of the `Scheduler._parse_submit_output` abstract method is
updated to allow to return an `ExitCode` instead of the job id. If an
exit code is returned, the engine will immediately terminate the
calculation job and assign the exit code to the node. This now allows
scheduler plugins to parse specific problems known to that particular
scheduler that are guaranteed to be unrecoverable and prevent the jobs
from unnecessarily going through the experimentall backoff mechanism and
eventually getting stuck.