Tests: Add test highlighting problem with mutation during iterall
There currently is a bug in the `PsqlDosBackend` that when the session
is commited during `QueryBuilder.iterall`, for example by creating a new
ORM instance, an exception is raised. This bug was addressed in the
commit 98e902bcc5d38363b5109c38cfae3e6073a3a23c, however, this caused
big problems that were not detected by the tests. Essentially, change
were no longer automatically committed and so changes performed in a
shell or by `verdi` commands would not get persisted to the database.

The introduced bug was released with `v2.2.0` but the responsible
commits have now been reverted. Here we add the test that demonstrates
the original bug that the reverted commits were trying to address as
described in https://github.com/aiidateam/aiida-core/issues/5802. For
now the test is skipped since it fails, but once the bug is properly
fixed, without breaking everything else, it should be enabled.