CLI: Prefix the `verdi data` subcommands with `core.`
The `verdi data` subcommands, which are pluginnable through the
`aiida.cmdline.data` entry point group, get their actual name from the
corresponding entry point name, not the name that is declared in the
code. However, the `cli_command_map` fixture gets the name from the code
and not the entry point name. A mismatch in these names will cause the
`run_cli_command` fixture to fail in the case of `use_subprocess=True`,
so therefore we correct the command names to match the entry point.