CLI: verdi export to a yaml file with keys from code class
fixes #3521

Add `verdi code export` command to export code from command line as a
ymal file.
This is mentioned in usability improvement as well as having a command to export the code and computer setup.
Keys of YAML file are read from the cli option of the corresponding code class.