fix: more flexible verdi status check
`verdi status` was checking for the presence of the daemon by the name
of the corresponding process. If the name was not as expected, it
deleted (!) the PID file.

On some systems, the name of the daemon process differed from the name
expected by AiiDA, causing `verdi status` to delete the PID file of a
correctly operating daemon.